+++ src/tickle/api/app.py
@@ -25,6 +25,7 @@ def _registerBlueprints(flask_app: flask.Flask):
 
     bp_version.register_blueprint(routes.bp_test, url_prefix='/test')
     bp_version.register_blueprint(routes.bp_watches, url_prefix='/watches')
+    bp_version.register_blueprint(routes.bp_internal, url_prefix='/internal')
 
     flask_app.register_blueprint(bp_version, url_prefix='/v1')
     
@@ -51,9 +52,9 @@ def _setConfigurations(flask_app: flask.Flask, selected_config: ConfigBase):
     pymysql.credentials.HOST     = selected_config.DB_HOST
 
 
+# Main logic
 app = flask.Flask(__name__)
 app_config = _getConfigurationClass(app.env)
-_setConfigurations(app, app_config)
 _registerBlueprints(app)
-
+_setConfigurations(app, app_config)
