+++ src/tickle/api/repository/watches.py
@@ -7,7 +7,6 @@ Watches repository
 """
 
 from __future__ import annotations
-import pymysql
 import pymysql.commands as sql_engine
 from pymysql.structs import DbOperationResult
 from tickle.common.domain import models
@@ -49,7 +48,8 @@ def insert(watch: models.Watch) -> DbOperationResult:
 
     return sql_engine.modify(SQL_INSERT, parms)
 
-
-
+#------------------------------------------------------
+# Get all the open watch records from the database
+#------------------------------------------------------
 def selectAllOpen() -> DbOperationResult:
     return sql_engine.selectAll(SQL_SELECT_ALL_OPEN)
