+++ src/tickle/api/routes/__init__.py
@@ -1,2 +1,3 @@
 from .test import bp_test as bp_test
 from .watches import bp_watches as bp_watches
+from .internal import bp_internal as bp_internal
