+++ src/tickle/api/services/__init__.py
@@ -1,4 +1,4 @@
 from . import email as email
 from . import watches as watches
-from . import inspector as inspector
+from . import audit as audit
 from . import tickerlib as tickerlib
